+++ src/contexts/index.js
@@ -1,0 +1,2 @@
+export { useCart, CartProvider } from "./CartProvider/CartProvider";
+export { useProduct, ProductProvider } from "./ProductProvider/ProductProvider";
